reorganized project folders
