+++ src/routes/covers/CoverOnly.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import {
     useGetCoverQuery,
-} from '../services/CoverServices';
+} from '../../store/covers/CoverServices';
 import ReactHtmlParser, { processNodes, convertNodeToElement, htmlparser2 } from 'react-html-parser';
