+++ src/routes/covers/Publisher.js
@@ -1,30 +1,66 @@
 import React, { useEffect, useState } from 'react';
-import { Header, Footer, H2 } from '../components';
+import { Header, Footer, H2 } from '../../components';
 import styled from 'styled-components';
-import { useNavigate, Link } from 'react-router-dom';
+import { useNavigate, Link, useParams } from 'react-router-dom';
 import {
-    useGetCoversQuery,
-} from '../services/CoverServices';
-import { api } from '../services/api';
+    useGetPublisherAllQuery,
+    useGetPublishersQuery,
+} from '../../store/covers/publishedServices';
+import { api } from '../../store/api/api';
 import ReactPaginate from 'react-paginate';
+import { useSelector } from 'react-redux';
 
-const Covers = () => {
+
+const Publisher = () => {
+    console.log('Rendered');
     const [page, setPage] = useState(0);
+    const [limit, setLimit] = useState(5);// used to set number of documents per page
     const navigate = useNavigate();
+    const user = useSelector((state) => state.auth.user);
 
-    const { data, isFetching } = useGetCoversQuery(page + 1);
-    // console.log('data', page);
-    const coverList = data?.data?.docs;
-    console.log(coverList);
+
+    const { data, isFetching } = useGetPublishersQuery({id: user._id, page: page + 1, limit: limit});
+    const coversAll= data?.data?.coversPublished;
 
     const [searchTerm, setSearchTerm] = useState('');
-    const results = coverList?.filter((cover) =>
-        cover.coverName.toLowerCase().includes(searchTerm),
-    );
     const handleChange = (e) => {
         setSearchTerm(e.target.value);
     };
+    // console.log(coversAll);
 
+    const [totalP, setTotalP] = useState([]);
+    useEffect(()=> {
+      totalMDB();
+    }, []);
+
+    const totalMDB = async () => {
+        fetch(`${process.env.REACT_APP_API_URL}/publishAll/` + user._id)
+        .then((response) => {
+            return response.json(); // << This is the problem
+        })
+        .then((responseData) => { // responseData = undefined
+            setTotalP(responseData.data.coversPublishedAll);
+            // setbio(responseData.json());
+            return responseData;
+        })
+        .catch(function(err) {
+          console.log(err);
+        });
+    };
+
+    const [filteredResults, setFilteredResults] = useState(coversAll);// Created to limit Rerendering of code
+
+    useEffect(() => {
+        const newFilteredResults = data?.data?.coversPublished.filter((cover) =>
+            cover.coverName.toLowerCase().includes(searchTerm),
+        );
+        setFilteredResults(newFilteredResults);
+
+        console.log('effect is firing');
+    }, [coversAll, searchTerm]);// run code only on change of these variables
+
+    // console.log(filteredResults);
+// console.log(Math.ceil(totalP/limit));
     return (
         <>
             <ScrollView>
@@ -40,10 +76,11 @@ const Covers = () => {
                         value={searchTerm}
                         onChange={handleChange}
                     />
-                    {
+
+                    {isFetching ? <div style={{ margin: '0 auto' }}><p>Loading...</p></div> :
                         <>
 
-                            {results?.map((item, index) => {
+                            {filteredResults?.map((item, index) => {
                                 return <div key={index} onClick={() => navigate(`/covers/${item._id}`)} style={{ padding: '10px', border: '2px solid gray', margin: '0 auto', marginBottom: '20px' }}>
                                     <p><b>Company:</b> {item.coverName}</p>
                                     <p><b>To:</b> {item.dear}</p>
@@ -54,7 +91,7 @@ const Covers = () => {
                                     initialPage={page}
                                     onPageChange={(page) => setPage(page.selected)}
                                     pageRangeDisplayed={5}
-                                    pageCount={data?.data?.totalPages}
+                                    pageCount={Math.ceil(totalP/limit)}
                                     activeClassName="active"
                                 />
                             </div>
@@ -67,7 +104,7 @@ const Covers = () => {
     );
 };
 
-export default Covers;
+export default Publisher;
 
 const ScrollView = styled.div`
     min-height: calc(100vh - 80px);
