+++ src/store/covers/publishedServices.js
@@ -1,4 +1,4 @@
-import { api } from './api';
+import { api } from '../api/api';
 
 export const PublishedService = api.injectEndpoints({
