+++ src/store/user/UserServices.js
@@ -1,4 +1,4 @@
-import { api } from './api';
+import { api } from '../api/api';
 
 export const UserServices = api.injectEndpoints({
